+++ NEWS.md
@@ -1,3 +1,11 @@
+# rflexscan 1.0.0
+
+* Added reference to the Journal of Statistical Software paper on rflexscan <doi:10.18637/jss.v099.i13>.
+
+# rflexscan 0.4.0
+
+* Allowed access to internal functions.
+
 # rflexscan 0.3.0
 
 * Updated help for `rflexscan` method.